Import PostsModule in AuthModule
i kinda feel like I wouldn't expect it to need to be nested under the
Auth module.. but I guess since it's using the auth guards and such -
same as the UsersModule, it needs to be nested a level down, rather than
being a sibling to the AuthModule.

would like to revisit this once I'm a bit further along to see if I'm
misunderstanding something tho for sure.