Create profile endpoint
since the client doesn't take the user id from the '/login' call, and
the token isn't stored in app state, when the client re-opens the app
(and the ServiceWorker still has the token), the client code needs a way
to check if the user is still authenticated from last time.

I figured the simplest way to do this is just send a request which would
fail without the token, and then when it succeeds, the client knows
there's a token ready for use and therefore there's a user session.

So yeah I felt that fetching the profile would be an appropriate call to
make - but without any user id, i can't fetch /user/:id to get the
profile, so I needed to set up a new route which just uses the user
which is found based on the token.